// src/components/quotes/types.ts
export type QuoteType = {
	id: number;
	last: string;
	lastColor: string;
	lowestAsk: string;
	highestBid: string;
	highestBidColor: string;
	percentChange: string;
	percentChangeColor: string;
	baseVolume: string;
	quoteVolume: string;
	isFrozen: string;
	postOnly: string;
	high24hr: string;
	low24hr: string;
};